requirements.txt
streamlit>=1.32.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
xgboost>=2.0.0
opencv-python-headless>=4.8.0
joblib>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
